requirements.txt
joblib==1.2.0
numpy==1.26.4
pandas==2.2.1
pillow==10.0.0